requirements.txt
pydantic-settings
python-dotenv
python-telegram-bot
